DrunkTester-main/requirements.txt
# Web Framework
flask==3.0.2
flask-restx==1.3.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-limiter==3.5.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.1
gunicorn==21.2.0
Werkzeug==3.0.1

# Machine Learning
torch==2.2.0
torchvision==0.17.0
transformers==4.37.2
sentencepiece==0.1.99
tokenizers==0.15.2
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0

# Audio Processing
SpeechRecognition==3.10.1
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80
face-recognition==1.3.0

# Database
SQLAlchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# API and Documentation
swagger-ui-bundle==1.0.0
apispec==6.3.0
apispec-webframeworks==0.5.2

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2024.1
tqdm==4.66.1
cachetools==5.3.2
python-magic==0.4.27

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0 